Split multiple components in CurrentEvent file into another files
