+++ client/app/components/conference_room/event/NoEvent.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const NoEvent = () => <p>No event currently in progress</p>;
+
+export default NoEvent;
